+++ src/VrokSymfonyAddonsBundle.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace Vrok\SymfonyAddons;
 
-use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
 
 class VrokSymfonyAddonsBundle extends AbstractBundle
 {
-    public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void
-    {
-        $loader = new YamlFileLoader($builder, new FileLocator(__DIR__.'/../config'));
-        $loader->load('services.yaml');
+    public function loadExtension(
+        array $config,
+        ContainerConfigurator $container,
+        ContainerBuilder $builder
+    ): void {
+        $container->import(__DIR__.'/../config/services.yaml');
     }
 }