+++ tests/Command/CronDailyCommandTest.php
@@ -30,7 +30,10 @@ class CronDailyCommandTest extends KernelTestCase
         $commandTester = new CommandTester($command);
         $commandTester->execute([]);
 
-        $this->assertSame(0, $commandTester->getStatusCode());
+        $this->assertSame(
+            CronDailyCommand::SUCCESS,
+            $commandTester->getStatusCode()
+        );
     }
 
     public function testService(): void