+++ tests/Command/CronHourlyCommandTest.php
@@ -30,7 +30,10 @@ class CronHourlyCommandTest extends KernelTestCase
         $commandTester = new CommandTester($command);
         $commandTester->execute([]);
 
-        $this->assertSame(0, $commandTester->getStatusCode());
+        $this->assertSame(
+            CronHourlyCommand::SUCCESS,
+            $commandTester->getStatusCode()
+        );
     }
 
     public function testService(): void