+++ tests/Command/CronMonthlyCommandTest.php
@@ -30,7 +30,10 @@ class CronMonthlyCommandTest extends KernelTestCase
         $commandTester = new CommandTester($command);
         $commandTester->execute([]);
 
-        $this->assertSame(0, $commandTester->getStatusCode());
+        $this->assertSame(
+            CronMonthlyCommand::SUCCESS,
+            $commandTester->getStatusCode()
+        );
     }
 
     public function testService(): void